add fuzzy matching capabilities
This tackles two main flaws:

The station to be played is picked from the top of the list (NodeList). Since this not always reflects the uttered station (eg a request for "Deutschlandfunk" plays the stream of "Deutschlandfunk Nova" as a result - while "Deutschlandfunk" would be an option) fuzzy matching is introduced to ensure the best match is played.

Secondly, a TTS or otherwise butchered utterance (like with the example above if it understands "Deutschlandfunk Nowa" (w)) would result in "station not found". This inconvenience is intercepted by utilizing the APIs fuzzy matching capability. In this case the request is performed twice with the top suggested station

(cherry picked from commit b95005b4621dbd3d93e052fd6974e852e1b98e56)